+++ @geogebra/solver-sdk/test/gm-action-test-filters.ts
@@ -23,6 +23,139 @@ export const ruleSkipList = [
   // TODO: adjust the solver to do x*x => x^2 instead of x^(1+1); GM does that
   // when multiplying with an implied ^1 power (e.g. x^6*x or x*x but not x^(1+2)*x)
   'General.RewriteProductOfPowersWithSameBase',
+
+  /* TODO: Failing: 41 of 41
+  Problems:
+  - [OK] use 'left-of' instead of 'onto' or 'inside-left'
+  - [OK] need to call finish_interaction or we'll be left with extra parentheses
+  - [XX] need to pick up all of the remaining product, not just the first factor
+  */
+  'Expand.DistributeMultiplicationOverSum',
+
+  /*
+  TODO: Look into this more
+  In (x*(x+1))/(x+1) instead of moving x+1:2 onto x+1 we should move /(x+1) onto *(x+1)
+  We are currently using PM.Group but that does not seem to select the right targets
+  */
+  'General.CancelDenominator',
+
+  /* In gm we can't bring two fraction to the same denominator using a single
+  action. Example: in 3+1/2+1/3 solver expects that by moving 2 around 1/3 we
+  would get 3+1*3/2*3+1*2/3*2, while we can only transform to 3 + 1/2 + 2/2*3 
+  */
+  'FractionArithmetic.BringToCommonDenominator',
+  /* I adjusted this one so that the action would at least happen, but GM does
+  not simplify the expression as much as solver: In 5*x+(-4)*x^2 when moving the
+  - in front of the product we get 5*x+-(4)*x^2 instead of 5*x-4*x^2*/
+  'General.MoveSignOfNegativeFactorOutOfProduct',
+  /* TODO: Look into this more
+    It seems that we need the parent of the actors/targets we are getting here, where should this adjustment be made?
+    Check createPathMap, rearrangeNegativeProducts for errors as the selectors from solver are mismatched
+  */
+  'General.SimplifyProductWithTwoNegativeFactors',
+
+  //  ===========================================================================  //
+  // The tests below were made after the adjustments to the GMActions in the Rules
+  //  ===========================================================================  //
+
+  /* TODO: Look into this more
+    It seems that we need the parent of the actors/targets we are getting here, where should this adjustment be made?
+    Check createPathMap, rearrangeNegativeProducts for errors as the selectors from solver are mismatched
+  */
+  'General.FactorMinusFromSum',
+  'General.RemoveRedundantPlusSign',
+
+  /* The expressions are not simplified the same way in Solver and GM:
+    - In case of sums: GM leaves the + resulting in expressions with '+-' which should be simplified to only '-'
+    - In case of powers: GM removes the parentheses around the negative element
+  */
+  'IntegerArithmetic.SimplifyOddPowerOfNegative',
+  /* Skip this for now, can't really figure out what's happening
+   */
+  'Expand.ExpandProductOfSumAndDifference',
+  /* The expressions are not simplified the same way in Solver and GM:
+    - Gm simplifies x*x to x^2
+    - Gm simplifies products e.g. 2*2*x*x to 4*x^2
+  */
+  'Expand.ApplyFoilMethod',
+  /* The expressions are not simplified the same way in Solver and GM:
+    - Gm simplifies powers automatically e.g. (2*x)^2 - 4*x^2; 3^2 vs 9; 2*x*2*x - 4*x^2
+  */
+  'Expand.ExpandBinomialSquaredUsingIdentity',
+  'Expand.ExpandTrinomialSquaredUsingIdentity',
+  'Expand.ExpandDoubleBrackets',
+  'Expand.ExpandBinomialCubedUsingIdentity',
+
+  /*
+    - Gm does not simplify x^1 to x
+    - Gm does not distribute all the components from the exponents if there are
+      multiple, only the selected one, e.g 2^(a+b+c) - moving c to the right
+      will yield 2^(a+b)*2^c instead of 2^a*2^b*2^c as solver would expect
+  */
+  'General.DistributeSumOfPowers',
+  /*
+  The behavior of GM seems to be inconsistent in case of simplifying double
+  negatives:
+  - Sometimes the second minus should be dragged around the first, but the
+    parentheses won't be simplified (this seems to be the case with expressions
+    like -(-x) where there are no other components)
+  - Sometimes the first minus should be tapped
+  TODO: Ask Erik about this
+  */
+  'General.SimplifyDoubleMinus',
+  /* In this case the GM action seems to be wrong, as it tries to drag the
+  constant from the right side right-of some expression on the left side, but
+  manually executing the action indicates the action should be dragging left-of
+  the inequality sign/ right-inside the target expression
+  */
+  'Inequalities.MoveConstantsToTheLeft',
+  'Equations.MoveConstantsToTheLeft',
+  /* In these cases it seems that we need to use right-inside There doesn't seem
+    to be a Position.RightInside, and I couldn't figure out how to use it in GM
+    even if I added the type in solver-engine.
+    These are closely related to the ones above
+  */
+  'Equations.MoveConstantsToTheRight',
+  'Equations.MoveVariablesToTheLeft',
+  'Equations.MoveVariablesToTheRight',
+  'Equations.MoveConstantsInVariablesToTheRight',
+  'Equations.MoveConstantsInVariablesToTheLeft',
+  /* Similar to the previous one, although in this case the problem seems to be
+  that the actor should be dragged 'right-inside'? */
+  'Inequalities.CancelCommonTermsOnBothSides',
+  'Inequalities.MoveVariablesToTheRight',
+  'Inequalities.MoveConstantsToTheRight',
+  'Inequalities.MoveVariablesToTheLeft',
+  /* In this case the entire group should be selected, but even if the selection
+  were right, we would have different result from solver, I couldn't find a way
+  to drag only the denominator to the opposite side */
+  'Inequalities.MultiplyByInverseCoefficientOfVariableAndFlipTheSign',
+  /* This throws an error: TypeError: Cannot read properties of undefined (reading 'is_group'),
+    but the action itself seems to be right, even so it's possible that gm would simplify this
+    expression more then solver would*/
+  'Inequalities.DivideByCoefficientOfVariable',
+  /* I haven't been able to fix this but the target should use PM.Group as
+  otherwise the multiplication seems to be executed on expressions with
+  parentheses */
+  'General.ReorderProduct',
+  /* Gm does not automatically simplify sqrt[0] to 0 */
+  'Equations.TakeRootOfBothSidesRHSIsZero',
+  /* Parentheses are not automatically simplified in GM */
+  'Equations.TakeRootOfBothSides',
+  'Equations.NegateBothSides',
+  'General.DistributePowerOfProduct',
+  /* Parentheses are automatically simplified in GM */
+  'General.SimplifyOddPowerOfNegative',
+  /*
+  TODO: Look into this more
+  In (x*(x+1))/(x+1) instead of moving x+1:2 onto x+1 we should move /(x+1) onto *(x+1)
+  We are currently using PM.Group but that does not seem to select the right targets
+  */
+  'Inequalities.MultiplyByInverseCoefficientOfVariable',
+  'General.RewriteFractionOfPowersWithSameBase',
+  'Equations.MultiplyByInverseCoefficientOfVariable',
+  /* GM does not automatically simplify fractions when dividing both sides*/
+  'Decimals.SolveLinearEquation',
 ];
 
 export const expressionTypeSkipList = [
@@ -34,6 +167,8 @@ export const expressionTypeSkipList = [
   'Inequation',
   // TODO: support recurring decimals in GM (?)
   'RecurringDecimal',
+  // Constraints are not yet supported
+  'ExpressionWithConstraint',
 ];
 
 export const expressionSkipList = [
@@ -46,6 +181,7 @@ export const expressionSkipList = [
   // multiplication: a-2*(-3) => a-(-6) ; -(-1*x)+1 ==> -(-x)+1
   // leading addend: (-1)+1 => -1+1
   // drag factors on top of each other: (-2)(-3) => 6
+  // exponents: (x^2)^-2 => x^(2*(-2)) not x^(2*-2)
   '5*x+(1-5)*x^2',
   '(1-2)*a*x+1/3*a*b*x+3/4*a*b*x',
   '4*x^2+(-6-6)*x+9',
@@ -62,6 +198,7 @@ export const expressionSkipList = [
   '(-0.2)*(-0.2)',
   '-(-1*sqrt[3]3)-(-sqrt[3]3-3*sqrt[3]3+5*sqrt[3]3)',
   '-(3-4)*sqrt[3]3-(-sqrt[3]3-3*sqrt[3]3+5*sqrt[3]3)',
+  '(5^2)^(-{1/2})',
 
   // TODO: adjust GM to set entire product to 0 when multiplying by 0
   'x=(-7±sqrt(7^2-4*1*0))/(2*1)',
@@ -70,136 +207,24 @@ export const expressionSkipList = [
   '1^(sqrt 2+1)',
   '0^(3/2)',
 
-  // TODO: Added in PLUT-670 where factors of 1 are now eliminated
-  'x=(-3±sqrt(9+(-4)*1+(-4)*(-sqrt[3]3)))/2',
-];
+  // TODO: GM turns (2*a^2)^3 into 2^3*a^(2*3), while the solver stacks the
+  // exponents first. Adjust either the Solver or GM.
+  '(3*a*b^2*c*x^4)^3',
+  '(2*x^2)^2+x^2+(-3)^2+2*2*x^2*x+2*x*(-3)+2*(-3)*2*x^2',
+  '(2^3*3)^(2/3)',
+  '(5+2*x^2)*(25-10*x^2+(2*x^2)^2)',
+  '24*sqrt[3]((2^3*3)^2)',
+  'sqrt[6]((2^3*3)^5)',
+  '2*(2^2*3)^(1/2)',
 
-export const ruleWhiteList = [
-  'Collecting.CombineTwoSimpleLikeTerms',
-  'Decimals.ConvertFractionWithPowerOfTenDenominatorToDecimal',
-  'Decimals.EvaluateDecimalAddition',
-  'Decimals.EvaluateDecimalPowerDirectly',
-  'Decimals.EvaluateDecimalProduct',
-  'Decimals.EvaluateDecimalSubtraction',
-  'FractionArithmetic.SimplifyFractionToInteger',
-  'General.EliminateOneInProduct',
-  'General.EliminateZeroInSum',
-  'General.EvaluateExpressionToThePowerOfZero',
-  'General.EvaluateOneToAnyPower',
-  'General.EvaluateProductContainingZero',
-  'General.EvaluateZeroToAPositivePower',
-  'General.NormalizeNegativeSignOfIntegerInSum',
-  'General.RemoveBracketProductInProduct',
-  'General.RemoveBracketSumInSum',
-  'General.RemoveUnitaryCoefficient',
-  'General.RewriteFractionOfPowersWithSameExponent',
-  'General.RewriteProductOfPowersWithSameBase',
-  'General.RewriteProductOfPowersWithSameExponent',
-  'General.SimplifyExpressionToThePowerOfOne',
-  'General.SimplifyFractionWithOneDenominator',
-  'General.SimplifyUnitFractionToOne',
-  'General.SimplifyZeroNumeratorFractionToZero',
-  'IntegerArithmetic.EvaluateIntegerAddition',
-  'IntegerArithmetic.EvaluateIntegerPowerDirectly',
-  'IntegerArithmetic.EvaluateIntegerProduct',
-  'IntegerArithmetic.EvaluateIntegerSubtraction',
-  'IntegerArithmetic.SimplifyEvenPowerOfNegative',
-  'Decimals.ExpandFractionToPowerOfTenDenominator',
-  'Decimals.MultiplyFractionOfDecimalsByPowerOfTen',
-  // This one is weird as it contains the following test: 9900*x=31101 -> 31101/9900 with the x disappearing into thin air
-  // I guess even if we find the reason gm actions are not being executed we will still need to work on this
-  // TODO: ask solver team about this
-  //'Decimals.SolveLinearEquation',
+  // TODO: this should work in GM, find out why it doesn't
+  '24+2*2*sqrt 6*3*sqrt 2+(3*sqrt 2)^2',
 
-  // Also, any test case containing equations is prone to this very same behavior
+  // TODO: usually the Solver and GM don't keep a 0 with inverse addends,
+  // but the Solver does so here - adjust?
+  '(6-6+1)*x^2+30-x^6',
 
-  'Equations.DivideByCoefficientOfVariable',
-  // 'Equations.FlipEquation',
-  // 'Equations.MoveConstantsToTheLeft',
-  // 'Equations.MoveConstantsToTheRight',
-
-  // These cases are pretty stupid, if you move a variable to the other side of an equation,
-  // on the original side it should be replaced by zero, not the variable minus itself.
-  // We can't expect GM to have this solution.
-  // - Abel
-  // 'Equations.MoveVariablesToTheLeft',
-
-  // 'Equations.MoveVariablesToTheRight',
-  // 'Equations.MultiplyByInverseCoefficientOfVariable',
-  // 'Equations.MultiplyEquationByLCD',
-  // 'Equations.NegateBothSides',
-  // 'Equations.TakeRootOfBothSides',
-  // 'Equations.TakeRootOfBothSidesRHSIsZero',
-
-  // In this test for some cases the \cdot is hidden and the double tap is used, except
-  // it is not specified where to double tap (the \cdot can't be selected),
-  // Workaround: select the parenthesis that follows the operator
-  // Related problem: GM simplifies the expression while Solver does not (due to this some tests will fail if uncommented)
-  // - Abel
-  // 'Expand.ApplyFoilMethod',
-
-  // In these cases the behavior of Solver seems to be too different to align them up,
-  // GM will always simplify the distributed expressions (e.g instead of 4*11 we have 44)
-  // - Abel
-  // 'Expand.DistributeMultiplicationOverSum',
-
-  'Expand.DistributeNegativeOverBracket',
-
-  // GM simplifies the expression further than Solver in some test cases
-  // (e.g (2*x)^3 - 8*x^3)
-  // 'Expand.ExpandBinomialCubedUsingIdentity',
-
-  // GM simplifies the expression further than Solver in some test cases
-  // 'Expand.ExpandBinomialSquaredUsingIdentity',
-
-  // 'Expand.ExpandDoubleBrackets',
-  // 'Expand.ExpandProductOfSumAndDifference',
-  // 'Expand.ExpandTrinomialSquaredUsingIdentity',
-  // 'FractionArithmetic.BringToCommonDenominator',
-  // 'FractionArithmetic.ConvertIntegerToFraction',
-  // 'General.AddClarifyingBracket',
-  // 'General.CancelAdditiveInverseElements',
-  // 'General.CancelCommonTerms',
-
-  // Most of these work fine, except some isolated cases where the action is not executed
-  // even though the selectors and action type are correct
-  // - Abel
-  // 'General.CancelDenominator',
-
-  // 'General.CancelRootIndexAndExponent',
-  // 'General.DistributePowerOfProduct',
-  // 'General.DistributeSumOfPowers',
-  'General.EvaluateZeroToThePowerOfZero',
-  // 'General.FactorMinusFromSum',
-  // 'General.FlipFractionUnderNegativePower',
-  // 'General.MoveSignOfNegativeFactorOutOfProduct',
-  // 'General.MultiplyExponentsUsingPowerRule',
-  // 'General.RemoveRedundantBracket',
-  // 'General.RemoveRedundantPlusSign',
-
-  // Same here, even though selectors are correct, the expressions are not moved.
-  // I tried to select the parents (as when dragging onto) but it didn't change anything
-  // 'General.ReorderProduct',
-
-  // Same here, even though selectors are correct, the expressions are not moved.
-  // 'General.RewriteFractionOfPowersWithSameBase',
-
-  // 'General.RewriteIntegerOrderRootAsPower',
-  'General.RewritePowerAsProduct',
-  'General.RewritePowerUnderRoot',
-  'General.RewriteProductOfPowersWithInverseBase',
-  // 'General.RewriteProductOfPowersWithInverseFractionBase',
-  // 'General.RewriteProductOfPowersWithNegatedExponent',
-
-  // Here GM does not simplify the parentheses by default, but the actions are executed correctly
-  // 'General.SimplifyDoubleMinus',
-
-  'General.SimplifyProductOfConjugates',
-  // 'General.SimplifyProductWithTwoNegativeFactors',
-  // 'General.SimplifyZeroDenominatorFractionToUndefined',
-  // 'Inequalities.CancelCommonTermsOnBothSides',
-  // 'Inequalities.MoveConstantsToTheLeft',
-  // 'Inequalities.MoveConstantsToTheRight',
-  // 'Inequalities.MoveVariablesToTheLeft',
-  // 'Inequalities.MoveVariablesToTheRight',
+  // TODO: fix GM bug with dragging factors across equal sign
+  '-3*x=-1',
+  '3.2*x=-1',
 ];